misplaced tests.internals tests
