+++ pandas/tests/frame/test_block_internals.py
@@ -104,52 +104,6 @@ class TestDataFrameBlockInternals:
         float_frame[float_frame > 1] = 2
         tm.assert_almost_equal(expected, float_frame.values)
 
-    def test_values_numeric_cols(self, float_frame):
-        float_frame["foo"] = "bar"
-
-        values = float_frame[["A", "B", "C", "D"]].values
-        assert values.dtype == np.float64
-
-    def test_values_lcd(self, mixed_float_frame, mixed_int_frame):
-
-        # mixed lcd
-        values = mixed_float_frame[["A", "B", "C", "D"]].values
-        assert values.dtype == np.float64
-
-        values = mixed_float_frame[["A", "B", "C"]].values
-        assert values.dtype == np.float32
-
-        values = mixed_float_frame[["C"]].values
-        assert values.dtype == np.float16
-
-        # GH 10364
-        # B uint64 forces float because there are other signed int types
-        values = mixed_int_frame[["A", "B", "C", "D"]].values
-        assert values.dtype == np.float64
-
-        values = mixed_int_frame[["A", "D"]].values
-        assert values.dtype == np.int64
-
-        # B uint64 forces float because there are other signed int types
-        values = mixed_int_frame[["A", "B", "C"]].values
-        assert values.dtype == np.float64
-
-        # as B and C are both unsigned, no forcing to float is needed
-        values = mixed_int_frame[["B", "C"]].values
-        assert values.dtype == np.uint64
-
-        values = mixed_int_frame[["A", "C"]].values
-        assert values.dtype == np.int32
-
-        values = mixed_int_frame[["C", "D"]].values
-        assert values.dtype == np.int64
-
-        values = mixed_int_frame[["A"]].values
-        assert values.dtype == np.int32
-
-        values = mixed_int_frame[["C"]].values
-        assert values.dtype == np.uint8
-
     def test_constructor_with_convert(self):
         # this is actually mostly a test of lib.maybe_convert_objects
         # #2845
@@ -299,19 +253,6 @@ class TestDataFrameBlockInternals:
         # invalid error message on dt inference
         if not compat.is_platform_windows():
             f("M8[ns]")
-
-    def test_equals_different_blocks(self):
-        # GH 9330
-        df0 = DataFrame({"A": ["x", "y"], "B": [1, 2], "C": ["w", "z"]})
-        df1 = df0.reset_index()[["A", "B", "C"]]
-        # this assert verifies that the above operations have
-        # induced a block rearrangement
-        assert df0._mgr.blocks[0].dtype != df1._mgr.blocks[0].dtype
-
-        # do the real tests
-        tm.assert_frame_equal(df0, df1)
-        assert df0.equals(df1)
-        assert df1.equals(df0)
 
     def test_copy_blocks(self, float_frame):
         # API/ENH 9607
@@ -468,88 +409,6 @@ class TestDataFrameBlockInternals:
         result = df._get_numeric_data()
         expected = df.loc[:, ["A", "C"]]
         tm.assert_frame_equal(result, expected)
-
-    def test_convert_objects(self, float_string_frame):
-
-        oops = float_string_frame.T.T
-        converted = oops._convert(datetime=True)
-        tm.assert_frame_equal(converted, float_string_frame)
-        assert converted["A"].dtype == np.float64
-
-        # force numeric conversion
-        float_string_frame["H"] = "1."
-        float_string_frame["I"] = "1"
-
-        # add in some items that will be nan
-        length = len(float_string_frame)
-        float_string_frame["J"] = "1."
-        float_string_frame["K"] = "1"
-        float_string_frame.loc[float_string_frame.index[0:5], ["J", "K"]] = "garbled"
-        converted = float_string_frame._convert(datetime=True, numeric=True)
-        assert converted["H"].dtype == "float64"
-        assert converted["I"].dtype == "int64"
-        assert converted["J"].dtype == "float64"
-        assert converted["K"].dtype == "float64"
-        assert len(converted["J"].dropna()) == length - 5
-        assert len(converted["K"].dropna()) == length - 5
-
-        # via astype
-        converted = float_string_frame.copy()
-        converted["H"] = converted["H"].astype("float64")
-        converted["I"] = converted["I"].astype("int64")
-        assert converted["H"].dtype == "float64"
-        assert converted["I"].dtype == "int64"
-
-        # via astype, but errors
-        converted = float_string_frame.copy()
-        with pytest.raises(ValueError, match="invalid literal"):
-            converted["H"].astype("int32")
-
-        # mixed in a single column
-        df = DataFrame(dict(s=Series([1, "na", 3, 4])))
-        result = df._convert(datetime=True, numeric=True)
-        expected = DataFrame(dict(s=Series([1, np.nan, 3, 4])))
-        tm.assert_frame_equal(result, expected)
-
-    def test_convert_objects_no_conversion(self):
-        mixed1 = DataFrame({"a": [1, 2, 3], "b": [4.0, 5, 6], "c": ["x", "y", "z"]})
-        mixed2 = mixed1._convert(datetime=True)
-        tm.assert_frame_equal(mixed1, mixed2)
-
-    def test_infer_objects(self):
-        # GH 11221
-        df = DataFrame(
-            {
-                "a": ["a", 1, 2, 3],
-                "b": ["b", 2.0, 3.0, 4.1],
-                "c": [
-                    "c",
-                    datetime(2016, 1, 1),
-                    datetime(2016, 1, 2),
-                    datetime(2016, 1, 3),
-                ],
-                "d": [1, 2, 3, "d"],
-            },
-            columns=["a", "b", "c", "d"],
-        )
-        df = df.iloc[1:].infer_objects()
-
-        assert df["a"].dtype == "int64"
-        assert df["b"].dtype == "float64"
-        assert df["c"].dtype == "M8[ns]"
-        assert df["d"].dtype == "object"
-
-        expected = DataFrame(
-            {
-                "a": [1, 2, 3],
-                "b": [2.0, 3.0, 4.1],
-                "c": [datetime(2016, 1, 1), datetime(2016, 1, 2), datetime(2016, 1, 3)],
-                "d": [2, 3, "d"],
-            },
-            columns=["a", "b", "c", "d"],
-        )
-        # reconstruct frame to verify inference is same
-        tm.assert_frame_equal(df.reset_index(drop=True), expected)
 
     def test_stale_cached_series_bug_473(self):
